+++ mizani/tests/test_breaks.py
@@ -8,7 +8,7 @@ import pytest
 
 from ..breaks import (mpl_breaks, log_breaks, minor_breaks,
                       trans_minor_breaks, date_breaks,
-                      timedelta_breaks)
+                      timedelta_breaks, extended_breaks)
 from ..transforms import trans
 
 
@@ -131,10 +131,18 @@ def test_timedelta_breaks():
     major = breaks(limits)
     minutes = [val.total_seconds()/60 for val in major]
     npt.assert_allclose(
-        minutes, [0, 2, 4, 6, 8, 10])
+        minutes, [0, 2, 4, 6, 8])
 
     # numpy
     x = [np.timedelta64(i*10, unit='D') for i in range(1, 10)]
     limits = min(x), max(x)
     with pytest.raises(ValueError):
         breaks(limits)
+
+
+def test_extended_breaks():
+    x = np.arange(100)
+    limits = min(x), max(x)
+    for n in (5, 7, 10, 13, 31):
+        breaks = extended_breaks(n=n)
+        assert len(breaks(limits)) <= n+1